optimize install script with zsh

+++ install/macos.sh
@@ -1,9 +1,12 @@
 #!/usr/bin/env bash
 
 brew upgrade
+export HOMEBREW_NO_AUTO_UPDATE=1
 brew install macvim
 brew install neovim
-brew install zsh
+if [ ! -f /bin/zsh ]; then
+    brew install zsh
+fi
 brew install curl-openssl
 brew install wget
 brew install llvm
